+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/PodTemplateUtils.java
@@ -349,6 +349,7 @@ public class PodTemplateUtils {
 
         podTemplate.setCustomWorkspaceVolumeEnabled(template.isCustomWorkspaceVolumeEnabled() ?
                                                     template.isCustomWorkspaceVolumeEnabled() : parent.isCustomWorkspaceVolumeEnabled());
+        podTemplate.setPodRetention(template.getPodRetention());
 
         podTemplate.setYaml(template.getYaml() == null ? parent.getYaml() : template.getYaml());