+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/KubernetesPipelineTest.java
@@ -398,6 +398,15 @@ public class KubernetesPipelineTest extends AbstractKubernetesPipelineTest {
         r.assertLogContains("hostPort: 10001", b);
         r.assertLogContains("- containerPort: 10002", b);
         r.assertLogContains("hostPort: 10002", b);
+
+    @Test
+    public void runInPodWithRetention() throws Exception {
+        deletePods(cloud.connect(), getLabels(cloud, this), false);
+        WorkflowJob p = r.jenkins.createProject(WorkflowJob.class, "pod with retention");
+        p.setDefinition(new CpsFlowDefinition(loadPipelineScript("runInPodWithRetention.groovy"), true));
+        WorkflowRun b = p.scheduleBuild2(0).waitForStart();
+        r.assertBuildStatusSuccess(r.waitForCompletion(b));
+        assertTrue(deletePods(cloud.connect(), getLabels(this), true));
     }
 
 }